+++ src/commands/open.rs
@@ -61,7 +61,7 @@ pub(crate) fn run(opt: OptCompeteOpen, ctx: crate::Context<'_>) -> anyhow::Resul
     let member = metadata.query_for_member(package.as_deref())?;
     let package_metadata = member.read_package_metadata()?;
     let cargo_compete_config =
-        crate::config::load_from_rel_path(&member.manifest_path, &package_metadata.config)?;
+        crate::config::load_from_rel_path(&member.manifest_path, &package_metadata.config, shell)?;
 
     let mut urls = vec![];
     let mut file_paths = vec![];