+++ src/commands/retrieve_submission_summaries.rs
@@ -61,7 +61,7 @@ pub(crate) fn run(
     let metadata = crate::project::cargo_metadata(&manifest_path, cwd)?;
     let member = metadata.query_for_member(package.as_deref())?;
     let package_metadata = member.read_package_metadata()?;
-    crate::config::load_from_rel_path(&member.manifest_path, &package_metadata.config)?;
+    crate::config::load_from_rel_path(&member.manifest_path, &package_metadata.config, shell)?;
 
     let mut atcoder_targets = indexset!();